EPWeb/requirements.txt
acme==0.31.0
appdirs==1.4.4
asgiref==3.2.7
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cssselect2==0.3.0
cymem==2.0.3
deplacy==1.4.1
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.8
django-sslserver==0.22
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.0.12
future==0.15.2
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
googletrans==2.4.0
graphviz==0.14.1
grpcio==1.30.0
httplib2==0.9.2
hyperlink==17.3.1
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
identify==1.4.15
idna==2.9
incremental==16.10.1
Jinja2==2.10
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.5.0
MarkupSafe==1.0
mock==2.0.0
murmurhash==1.0.2
ndg-httpsclient==0.4.4
netifaces==0.10.4
nltk==3.4.5
nodeenv==1.3.5
numpy==1.19.1
oauthlib==2.0.6
packaging==20.3
PAM==0.4.2
pandas==1.1.0
parsedatetime==2.4
pbr==3.1.1
pgi==0.0.11.2
Pillow==7.2.0
plac==1.1.3
pluggy==0.13.1
pre-commit==2.4.0
preshed==3.0.2
protobuf==3.12.2
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==5.3.1
reportlab==3.5.44
requests==2.23.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rsa==4.6
SecretStorage==2.3.1
service-identity==16.0.0
six==1.14.0
spacy==2.3.2
spacy-conll==2.0.0
sqlparse==0.3.1
srsly==1.0.2
ssh-import-id==5.7
stanza==1.0.1
supervisor==4.1.0
svglib==1.0.0
systemd-python==234
thinc==7.4.1
tinycss2==1.0.2
toml==0.10.0
torch==1.6.0
tox==3.15.0
tqdm==4.48.2
translate==3.5.0
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
uWSGI==2.0.18
virtualenv==20.0.20
wasabi==0.7.1
webencodings==0.5.1
websocket-client==0.48.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
